netwerk/test/mochitests/test_android_content_uri.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for content:// URI</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<pre id="test">
<script class="testbody" type="text/javascript">
add_task(function() {
  let content = SpecialPowers.Cc["@mozilla.org/network/android-content-input-stream;1"]
                  .createInstance(SpecialPowers.Ci.nsIAndroidContentInputStream);
  let uri = SpecialPowers.Services.io
              .newURI("content://org.mozilla.geckoview.test_runner.provider/blob");
  content.init(uri);

  let sis = SpecialPowers.Cc["@mozilla.org/scriptableinputstream;1"]
              .createInstance(SpecialPowers.Ci.nsIScriptableInputStream);
  sis.init(content);

  is(sis.read(4), "ABCD", "data is valid");
});
</script>
</pre>
</body>
</html>
